vibe3d/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jsonschema>=4.20.0
websockets>=12.0
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyinstaller>=6.0.0
